Refactor to orchestate dataset initialization and loading from main module 'app.py'.

+++ init_db.py
@@ -1,4 +1,6 @@
+import os
 import pandas as pd
+
 def download_data(url):
     df = pd.read_csv(url)
     return df
@@ -13,12 +15,17 @@ def prepare_data(df):
 
 def init_db():
     url = 'https://raw.githubusercontent.com/plotly/Figure-Friday/refs/heads/main/2024/week-46/PDO_wine_data_IT_FR.csv'
-    print('Downloading data...')
-    raw_df = download_data(url)
-    print('Preparing data...')
-    prepared_df = prepare_data(raw_df)
     file_path = 'data/prepared_data.csv'
-    prepared_df.to_csv(file_path, index=False)
-    print(f'Data prepared and saved to {file_path}.')
+
+    if not os.path.exists('data'):
+        os.makedirs('data')
+    if not os.path.exists(file_path):
+        print('Downloading data...')
+        raw_df = download_data(url)
+        print('Preparing data...')
+        prepared_df = prepare_data(raw_df)
+        prepared_df.to_csv(file_path, index=False)
+        print(f'Data prepared and saved to {file_path}.')
+
 if __name__ == '__main__':
     init_db()
